+++ android_library/feedbacklibrary/src/main/java/ch/uzh/supersede/feedbacklibrary/activities/FeedbackSettingsActivity.java
@@ -20,6 +20,7 @@ import ch.uzh.supersede.feedbacklibrary.utils.ServiceUtility;
 import static ch.uzh.supersede.feedbacklibrary.utils.Constants.*;
 import static ch.uzh.supersede.feedbacklibrary.utils.Enums.SETTINGS_VIEW;
 import static ch.uzh.supersede.feedbacklibrary.utils.Enums.SETTINGS_VIEW.*;
+import static ch.uzh.supersede.feedbacklibrary.utils.PermissionUtility.USER_LEVEL.ADVANCED;
 
 
 @SuppressWarnings("squid:MaximumInheritanceDepth")
@@ -89,6 +90,9 @@ public class FeedbackSettingsActivity extends AbstractBaseActivity implements IF
                 getView(R.id.settings_layout_color_3, LinearLayout.class),
                 getView(R.id.settings_layout_color_4, LinearLayout.class));
         invokeVersionControl(5, R.id.settings_toggle_feature_3);
+        if (!ADVANCED.check(this)) {
+            disableViews(enableNotificationsToggle);
+        }
         onPostCreate();
     }
