+++ android_library/feedbacklibrary/src/main/java/ch/uzh/supersede/feedbacklibrary/utils/DefaultConfiguration.java
@@ -1,12 +1,6 @@
 package ch.uzh.supersede.feedbacklibrary.utils;
 
-import android.app.Activity;
-import android.content.Intent;
-
-import ch.uzh.supersede.feedbacklibrary.beans.LocalConfigurationBean;
 import ch.uzh.supersede.feedbacklibrary.entrypoint.*;
-
-import static ch.uzh.supersede.feedbacklibrary.utils.Constants.EXTRA_KEY_APPLICATION_CONFIGURATION;
 
 public class DefaultConfiguration implements
         IFeedbackBehaviorConfiguration,
@@ -202,13 +196,4 @@ public class DefaultConfiguration implements
         return "https://platform.supersede.eu:8443/";
     }
 
-    public static class DefaultConfigurationUtility {
-        public static LocalConfigurationBean getConfigurationFromActivity(Activity activity) {
-            return getConfigurationFromIntent(activity.getIntent());
-        }
-
-        public static LocalConfigurationBean getConfigurationFromIntent(Intent intent) {
-            return (LocalConfigurationBean) intent.getSerializableExtra(EXTRA_KEY_APPLICATION_CONFIGURATION);
-        }
-    }
 }
